(Hybrid-)GGA is supported
The bug that confused me lied in ambiguity of the book "Density Functional Theory in Quantum Chemistry (Takao Tsuneda)". Eq-4.15 applied only to LDA, not all kinds of DFs. The correct formulae for GGA and so on is found in "https://www.ncbi.nlm.nih.gov/pmc/articles/PMC7179435/".

+++ DensityFunctional.cpp
@@ -71,7 +71,7 @@ void ReadDF(std::string df,int & x,int & c,double & hf,char & approx,const bool 
 	}
 }
 
-void getEVxc(int id,double * ds,double * cgs,long int ngrids,double * es,double * vrs,double * vss){
+void getEVxc(int id,double * ds,double * cgs,int ngrids,double * es,double * vrs,double * vss){
         xc_func_type func;
         xc_func_init(&func,id,XC_UNPOLARIZED);
 	switch(func.info->family){
