+++ HartreeFock.h
@@ -1,6 +1,6 @@
 double RKS(int nele,EigenMatrix overlap,EigenMatrix hcore,
             double * repulsion,short int * indices,long int n2integrals,
-            int dfxid,int dfcid,long int ngrids,double * gridweights,
+            int dfxid,int dfcid,int ngrids,double * gridweights,
             double * gridaos,
             double * gridao1xs,double * gridao1ys,double * gridao1zs,
             EigenVector & orbitalenergies,EigenMatrix & coefficients,EigenMatrix & density,
